Bug fix; comments and home components
